Add the possibility to initialize a test suite with a static init() method rather than using the main() method.
It is not sufficient to setup the static AllTests class via the main method.
This only allows to directly execute the AllTests.php file by parsing it
with php. Using phpunit is not possible as this will only use the suite()
method to grab the test suite.

It would still be better to use late static binding here but the added
method should be enough to allow running the tests in both ways.

The AllTests.php in Kolab_Session demonstrates how to use the init() method.